% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package.R
\docType{package}
\name{Covid19EstimationIl6JakInhibitors}
\alias{Covid19EstimationIl6JakInhibitors}
\alias{Covid19EstimationIl6JakInhibitors-package}
\title{Covid19EstimationIl6JakInhibitors}
\description{
Covid19EstimationIl6JakInhibitors
}
